feat: improved ui code
