+++ client/src/components/form-builder/form-section.tsx
@@ -1,0 +1,102 @@
+import React from 'react';
+import { Delete } from '@mui/icons-material';
+import {
+  Box,
+  Button,
+  Card,
+  CardContent,
+  Grid,
+  IconButton,
+  TextField,
+  Typography,
+} from '@mui/material';
+import { Controller, useFieldArray, useFormContext } from 'react-hook-form';
+import { DEFAULT_FIELD_VAL } from './const';
+import { FormField } from './form-field';
+
+interface FormSectionProps {
+  index: number;
+  onRemove: () => void;
+}
+
+export const FormSection = ({ index, onRemove }: FormSectionProps) => {
+  const { control } = useFormContext();
+  const { fields, append } = useFieldArray({
+    control,
+    name: `sections.${index}.fields`,
+  });
+
+  return (
+    <Card sx={{ mb: 4 }}>
+      <CardContent>
+        <Grid container spacing={2}>
+          <Grid item xs={12} sx={{ display: 'flex' }}>
+            <Controller
+              name={`sections.${index}.name`}
+              control={control}
+              rules={{
+                required: true,
+                minLength: 1,
+              }}
+              render={({
+                field: { onChange, value },
+                fieldState: { invalid },
+              }) => (
+                <TextField
+                  fullWidth
+                  label="Title"
+                  variant="outlined"
+                  onChange={onChange}
+                  value={value}
+                  error={invalid}
+                />
+              )}
+            />
+            <IconButton
+              aria-label="delete this section"
+              sx={{ ml: 1 }}
+              onClick={() => onRemove()}
+            >
+              <Delete />
+            </IconButton>
+          </Grid>
+          <Grid item xs={12}>
+            <Controller
+              name={`sections.${index}.description`}
+              control={control}
+              rules={{
+                required: true,
+                minLength: 1,
+              }}
+              render={({
+                field: { onChange, value },
+                fieldState: { invalid },
+              }) => (
+                <TextField
+                  fullWidth
+                  multiline
+                  label="Description (optional)"
+                  variant="outlined"
+                  onChange={onChange}
+                  value={value}
+                  error={invalid}
+                />
+              )}
+            />
+          </Grid>
+        </Grid>
+        <Typography variant="h5" sx={{ mt: 3, mb: 2 }}>
+          Questions
+        </Typography>
+        {fields.map((f, i) => (
+          <FormField key={f.id} index={i} sectionIndex={index} />
+        ))}
+        <Box sx={{ mt: 2 }}>
+          <Button fullWidth onClick={() => append(DEFAULT_FIELD_VAL)}>
+            Add question to section
+          </Button>
+        </Box>
+      </CardContent>
+    </Card>
+  );
+};
